+++ UI/DebugOverlay.cpp
@@ -97,9 +97,9 @@ static void DrawControlDebug(UIContext *ctx, const ControlMapper &mapper, const 
 
 static void DrawFrameTimes(UIContext *ctx, const Bounds &bounds) {
 	FontID ubuntu24("UBUNTU24");
-	double *sleepHistory;
+	float *sleepHistory;
 	int valid, pos;
-	double *history = __DisplayGetFrameTimes(&valid, &pos, &sleepHistory);
+	float *history = __DisplayGetFrameTimes(&valid, &pos, &sleepHistory);
 	int scale = 7000;
 	int width = 600;
